% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rasch.R
\docType{class}
\name{Rasch-class}
\alias{Rasch-class}
\title{A Rasch object}
\description{
Object of class \code{Rasch} needed to use functions in \code{raschPack}
}
\details{
An object of the class `Rasch' has the following slots:
\itemize{
\item \code{name} Character vector containing name of student
\item \code{difficulty} Numeric vector containing known difficulty parameters of each question
\item \code{correct_answers} Numeric vector indicating whether student answered correctly (1) or not (0)
}
}
\author{
David R. Miller: \email{drmiller@wustl.edu}
}
