% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschEAP.R
\name{raschEAP}
\alias{raschEAP}
\title{Calculating the expected a posteriori ability parameter value}
\usage{
raschEAP(raschObj, lower = -6, upper = 6, ...)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{lower}{The lower limit of integration; default is -6}

\item{upper}{The upper limit of integration; default is -6}
}
\value{
The expected a posteriori ability parameter
}
\description{
Calculates the expected a psoteriori value for the ability parameter for a student given
their answers and the difficulty of the questions by integrating over the product of the
likelihood and the prior within set limits of integration
}
\author{
David R. Miller <\email{drmiller@wustl.edu}>
}
