% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschInfo.R
\name{raschInfo}
\alias{raschInfo}
\title{Calculating the Fisher's Information for each question}
\usage{
raschInfo(raschObj, theta, ...)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A proposed ability parameter}
}
\value{
A vector of the Fisher's Information for each question
}
\description{
To calculate the amount of Fisher's Information we obtain for a student with a given
ability parameter for each question item, we evaluate the negative second derivative of
the likelihood.
}
\author{
David R. Miller <\email{drmiller@wustl.edu}>
}
