% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschLik.R
\name{raschLik}
\alias{raschLik}
\title{Calculating the likelihood of a proposed ability parameter}
\usage{
raschLik(raschObj, theta, ...)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A proposed ability parameter}
}
\value{
The likelihood of the proposed theta
}
\description{
Calculates the likelihood of a proposed ability parameter given the responses we observe
}
\author{
David R. Miller <\email{drmiller@wustl.edu}>
}
