% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschPack-package.R
\docType{package}
\name{raschPack}
\alias{raschPack}
\alias{raschPack-package}
\title{raschPack}
\description{
The raschPack package calculates ability parameters using a Rasch model
}
\examples{


}
\author{
David R. Miller: \email{drmiller@wustl.edu}
}
