% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschPlot.R
\name{raschPlot}
\alias{raschPlot}
\title{Plots the item characteristic curves for Rasch objects}
\usage{
raschPlot(raschObj, ower_bound = -10, upper_bound = 10,
  questions = c(1:4), theta_EAP = TRUE, ...)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{upper_bound}{The largest value of theta to be included in the plot (default is 10)}

\item{questions}{The question items to be plotted.  By default, the function plots the
first four item characteristic curves on separate plots}

\item{theta_EAP}{Whether the plots should include the students EAP (default is true)}

\item{lower_bound}{The smallest value of theta to be included in the plot (default is -10)}
}
\value{
Separate plots for each question requested by the user
}
\description{
Produces plots of item characteristic curves for some questions answered by students of
class Rasch.  Enables the user to specify whether the EAP should also be plotted.
}
\author{
David R. Miller <\email{drmiller@wustl.edu}>
}
