% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschPrior.R
\name{raschPrior}
\alias{raschPrior}
\title{Calculating a prior for Rasch estimation}
\usage{
raschPrior(theta, ...)
}
\arguments{
\item{theta}{A proposed ability parameter}
}
\value{
The prior probability of the proposed ability parameter
}
\description{
Calculates the prior for Rasch estimation given a proposed ability parameter.  The prior
probability distribution is fixed as a normal distribution with mean 0 and standard
deviation 3
}
\author{
David R. Miller <\email{drmiller@wustl.edu}>
}
