% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raschProb.R
\name{raschProb}
\alias{raschProb}
\title{Evaluating probability of correct responses}
\usage{
raschProb(raschObj, theta, ...)
}
\arguments{
\item{raschObj}{An object of class Rasch}

\item{theta}{A proposed ability parameter}
}
\value{
A list with the elements
 \item{prob_corr}{The probability of a correct answer given the proposed ability parameter}
 \item{prob_obs}{The probability of the observed answers given the proposed ability parameter}
}
\description{
Determines the probability of a correct response given an ability parameter, and the probability of the observed responses
}
\author{
David R. Miller <\email{drmiller@wustl.edu}>
}
